refactor NBT
remove C++11 warnings (override)
move Tag* code into own sub files
